requirements.txt
Django==3.2.3
python-dotenv==1.0.0
djangorestframework==3.12.4
djoser==2.2.0
pillow==10.0.0
django-colorfield==0.7.2